IA/app/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.1.0
whisper>=1.0.0
moviepy>=1.0.0
psycopg2-binary>=2.9.0
matplotlib>=3.0.0
langchain>=0.0.200
langchain-openai>=0.0.1
openai-whisper>=1.0
ffmpeg-python>=0.2.0
torch>=2.0.0
torchaudio>=2.0.0